build: create dockerfile (#58)
**What this PR does**:
This pr adds dockerfile to api service.

Some notes before code review:
I decided to use multi-stage builds technique because it is very popular technique when you are using golang, multi-stage builds make it easier to borrow automations and artifacts from other images, and allows you to build a binary in one image and copy only it to the final image.

I tested two different base images: `alpine` `scratch`.
`scratch` - is an empty image, so once the container running, we can’t exec into the container because it doesn’t even have a shell command.
`alpine` - is based on the [Alpine Linux Project](https://alpinelinux.org/), which is an operating system that was built specifically for use inside of containers.

Even though alpine creates a bigger image (18mb vs 11mb) I chose it, Why? because the size is small and it still has several commands and a shell command so we can docker exec into the container and access it. If we use the scratch base image, it will be hard for us to debug our running container because we can’t exec into it.

About the location of the dockerfile?
I chose to place it inside the cmd/api, Why? because i want to be able support several different images in the future for example (Query service, ingestion service), in addition i looked some popular projects such as [Otel](https://github.com/open-telemetry/opentelemetry-collector-contrib/tree/main/cmd/tracegen).

**Which issue(s) this PR fixes**:
Fixes #48 

**Checklist**
- [ ] Tests updated
- [x] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`