fix: use yarn instead of npm (#113)
**What this PR does**:
The project started with yarn as a package manager and lately npm support was added (probably by mistake).
Using multiple package managers can cause issues and is not recommended.
The purpose of this PR is to update our codebase to use a single package manager by:
1. Removing npm's package-lock.json file
2. Updating scripts to use yarn instead of npm

I don't really care about the yarn vs npm wars at the moment, and I don't mind changing it in the future or even now. But I don't want other PR's to be merged in the meantime and cause conflicts for other developers.
Feel free to close it and replace it with npm if you think it is a better choice.