feat: create logs package (#137)
**What this PR does**:
Since the logger is going to be initialized from multiple main functions (per each service - #133), it makes sense to create a package to share that logic.
This PR also does some small logical changes, for example, the logger is initialized from the application config instead of reading environment variables on its own. It is much cleaner and easier to understand when the config serves as the single source of truth for app configurations.

**Which issue(s) this PR fixes**:
Fixes #134 

**Checklist**

- [x] Tests updated
- [x] Documentation added