feat(collector): add OTLP traces receiver (#168)
**What this PR does**:
This PR adds the first part in the traces ingestion architecture:
OTLP receiver (gRPC/HTTP) -> Span formatting (flattening/adding fields) -> DB exporters (e.g. Elastic)

This is the basic implementation of the OTLP receiver and supports both gRPC and HTTP traffic.
The main idea here is that the we could create and maintain our own gRPC/HTTP servers with grpc-gateway or a similar package and handle all of the infrastructure ourselves (e.g. maintaining proto files, routing, logging, etc.).
Instead, this code simply uses the existing OTLP receiver functionality from the OTEL collector package. This ensures that we have 100% compatibility with OTEL, future changes to the spec (e.g. how the collector sends traces) requires just a simple package upgrade.
The main idea for how to execute this kind of integration is taken from Jaeger which also uses this method in their own OTLP receiver.

This is the basic implementation, some other features will be added in the future:
- Unit tests (due to time constraints, it wasn't be added in this PR)
- TLS settings support
- Configurable gRPC message limitation settings (e.g. message size)
- TracesProcessor might use a different interface, depending on the work in other areas of the traces ingestion

![Screen Shot 2022-09-21 at 11 25 06](https://user-images.githubusercontent.com/37577482/191454394-16e8bdb8-4ccb-47b8-b1a8-b82f230d1c77.png)


**Which issue(s) this PR fixes**:
Part of #96 (missing tests - might use a new issue for that)

**Checklist**

- [ ] Tests updated
- [x] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`