feat(trace-view): Add resize handle to timeline component (#798)
**What this PR does:**
Implements  a handle that would allow dragging the timeline top edge and expand the view area.

**Which issue(s) this PR fixes:**
Fixes #572 

**Additional Information:**

****Library choice****
I've tried three different libraries:

1. [Splitter](https://github.com/devbookhq/splitter/blob/master/example/src/ScrollableChildren/index.tsx)
2. [react-split-pane](https://github.com/tomkp/react-split-pane)
3. [Re-Flex](https://github.com/leefsmp/Re-Flex)

I used [Splitter](https://github.com/devbookhq/splitter/blob/master/example/src/ScrollableChildren/index.tsx) because it has minimal code addition upon implementation, has a lot of examples and was the easiest to use between the three.

****Design****
After discussing the design with @adva-cisco it was decided that when hovering the timeline's top border we will add a box-shadow so it will be clearer that it is draggable.