fix(sqlite): write empty event/link (#1348)
## What this PR does:
Writes an empty event/link for spans with no events/links. That's required in order for the full functionality of the SQLite spans storage to work properly (see #1317 ). Without that fix, spans with no events/links won't be returned for queries with specific filters. 
For example, querying spans with "spans.events.X not contains 'ABC'" filter will cause spans with no events to be dropped and not returned (since we use INTERSECT).  
In the future, the schema should be changed to fit the expected queries and that fix won't be needed.
Alternatively, we might change INTERSECT and use another solution in order to query the spans/tag values.

## Which issue(s) this PR fixes:
Fixes #1082