fix: run e2e tests on latest changes (#1479)
## What this PR does:
Currently, the e2e tests are running against the latest published image. But the PR changes are yet to be published as an image when the e2e tests run. This leads to an issue that the tests always run on the previously published image instead of the newest changes introduced in the PR.
This PR fixes the issue by building a new image locally and testing against it instead of pulling the latest published image.