fix: remove dynamic branch badges (#1582)
**Issue:**
The e2e tests action badge shows an incorrect status (failed instead of success).

**Reason:**
This is because a Github action badge must point to a specific branch (default one if no query param is supplied).
In our case, the e2e tests are running on a dynamic branch name created by Mergify on each PR, which can't be referred to in the badge URL.

**Solution:**
It is not that important. Let's remove it for now instead of showing an incorrect result.
Also, the merge-queue badge suffers from the same issue (shows success because it probably refers to some tests we've done on the main branch in the past). So I've also changed it to the [Mergify badge](https://docs.mergify.com/badge/).