docs(docker-compose): use latest instead of versioned image (#1584)
We can't keep updating the docs with each new release, and it doesn't make much sense since the `:latest` tag should always point to the latest stable release (excluding `:demo` releases).
It makes much more sense to use the `:latest` tag in the docs with a note of how to use it in production deployment.